Move email field and selector components to child directory
